adk-line-chatbot-template/requirements.txt
google-adk==1.27.1
flask==3.1.3
line-bot-sdk==3.22.0
python-dotenv==1.2.2
requests==2.32.5
